+++ modules/dnn/src/torch/THFile.h
@@ -27,7 +27,7 @@ TH_API unsigned char THFile_readByteScalar(THFile *self);
 TH_API char THFile_readCharScalar(THFile *self);
 TH_API short THFile_readShortScalar(THFile *self);
 TH_API int THFile_readIntScalar(THFile *self);
-TH_API long THFile_readLongScalar(THFile *self);
+TH_API long long THFile_readLongScalar(THFile *self);
 TH_API float THFile_readFloatScalar(THFile *self);
 TH_API double THFile_readDoubleScalar(THFile *self);
 
@@ -35,7 +35,7 @@ TH_API void THFile_writeByteScalar(THFile *self, unsigned char scalar);
 TH_API void THFile_writeCharScalar(THFile *self, char scalar);
 TH_API void THFile_writeShortScalar(THFile *self, short scalar);
 TH_API void THFile_writeIntScalar(THFile *self, int scalar);
-TH_API void THFile_writeLongScalar(THFile *self, long scalar);
+TH_API void THFile_writeLongScalar(THFile *self, long long scalar);
 TH_API void THFile_writeFloatScalar(THFile *self, float scalar);
 TH_API void THFile_writeDoubleScalar(THFile *self, double scalar);
 
@@ -63,7 +63,7 @@ TH_API long THFile_readByteRaw(THFile *self, unsigned char *data, long n);
 TH_API long THFile_readCharRaw(THFile *self, char *data, long n);
 TH_API long THFile_readShortRaw(THFile *self, short *data, long n);
 TH_API long THFile_readIntRaw(THFile *self, int *data, long n);
-TH_API long THFile_readLongRaw(THFile *self, long *data, long n);
+TH_API long THFile_readLongRaw(THFile *self, long long *data, long n);
 TH_API long THFile_readFloatRaw(THFile *self, float *data, long n);
 TH_API long THFile_readDoubleRaw(THFile *self, double *data, long n);
 TH_API long THFile_readStringRaw(THFile *self, const char *format, char **str_); /* you must deallocate str_ */
@@ -72,7 +72,7 @@ TH_API long THFile_writeByteRaw(THFile *self, unsigned char *data, long n);
 TH_API long THFile_writeCharRaw(THFile *self, char *data, long n);
 TH_API long THFile_writeShortRaw(THFile *self, short *data, long n);
 TH_API long THFile_writeIntRaw(THFile *self, int *data, long n);
-TH_API long THFile_writeLongRaw(THFile *self, long *data, long n);
+TH_API long THFile_writeLongRaw(THFile *self, long long *data, long n);
 TH_API long THFile_writeFloatRaw(THFile *self, float *data, long n);
 TH_API long THFile_writeDoubleRaw(THFile *self, double *data, long n);
 TH_API long THFile_writeStringRaw(THFile *self, const char *str, long size);
